// src/app/modules/core/interfaces/viajes.interface.ts
export interface Viajes{
    id: string;
    USUARIO_ID:string;
    numeroTripulacion: string;
    destino: string;
    numeroDias: string;
    experiencia: string;
    precio: string;
    incluye: string;
    embarcacion_id: string;
    urlFoto: string;
    origen: string;
    puertoMarina: string;
    numeroPasajeros: any;
    preciodesc?: any;
    capitan_barco?: any;
    embarcacionNombre?:any;
    lat?:any;
    lng?:any;


}